Seperate UI related methods in UI class, previously IOHelper
