--- src/main/java/Main.java
@@ -1,6 +1,0 @@
-public class Main {
-    public static void main(String[] args) {
-        Duke duke = new Duke();
-        duke.run();
-    }
-}
